+++ library/data-store/src/main/java/io/mehow/laboratory/datastore/DataStoreFeatureStorage.kt
@@ -1,14 +1,11 @@
 package io.mehow.laboratory.datastore
 
-import android.content.Context
 import androidx.datastore.core.DataStore
-import androidx.datastore.core.DataStoreFactory
 import io.mehow.laboratory.Feature
 import io.mehow.laboratory.FeatureStorage
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import okio.IOException
-import java.io.File
 
 internal class DataStoreFeatureStorage(
   private val dataStore: DataStore<FeatureFlags>,
@@ -46,24 +43,4 @@ internal class DataStoreFeatureStorage(
  */
 public fun FeatureStorage.Companion.dataStore(dataStore: DataStore<FeatureFlags>): FeatureStorage {
   return DataStoreFeatureStorage(dataStore)
-}
-
-/**
- * Creates a [FeatureStorage] that is backed by [DataStore] with a file taken from [fileProvider].
- */
-@Deprecated(
-    "This function will be removed in 1.0.0. Use FeatureStorage.dataStore(DataStore) instead.",
-)
-public fun FeatureStorage.Companion.dataStore(fileProvider: () -> File): FeatureStorage {
-  return dataStore(DataStoreFactory.create(FeatureFlagsSerializer, produceFile = fileProvider))
-}
-
-/**
- * Creates a [FeatureStorage] that is backed by [DataStore] with a file in an apps default directory.
- */
-@Deprecated(
-    "This function will be removed in 1.0.0. Use FeatureStorage.dataStore(DataStore) instead.",
-)
-public fun FeatureStorage.Companion.dataStore(context: Context, fileName: String): FeatureStorage {
-  return dataStore(DataStoreFactory.create(FeatureFlagsSerializer) { File(context.filesDir, "datastore/$fileName") })
 }