+++ library/docs/changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade Android target and compile SDK to 31.
 - Upgrade to LifecycleViewmodelKtx `2.4.0`.
 - Upgrade to KotlinPoet `1.10.2`.
+- Upgrade to Wire `3.7.1`.
 
 ## [0.14.0] - 2021-10-11